+++ Terminal.Gui/View/Layout/Dimension.cs
@@ -1,0 +1,22 @@
+namespace Terminal.Gui;
+
+/// <summary>
+///     Indicates the dimension for <see cref="Dim"/> operations.
+/// </summary>
+public enum Dimension
+{
+    /// <summary>
+    ///     No dimension specified.
+    /// </summary>
+    None = 0,
+
+    /// <summary>
+    ///     The height dimension.
+    /// </summary>
+    Height = 1,
+
+    /// <summary>
+    ///     The width dimension.
+    /// </summary>
+    Width = 2
+}
