+++ Terminal.Gui/View/Layout/Pos.cs
@@ -1,30 +1,5 @@
+#nullable enable
 namespace Terminal.Gui;
-
-/// <summary>
-///     Indicates the side for <see cref="Pos"/> operations.
-/// </summary>
-public enum Side
-{
-    /// <summary>
-    ///     The left (X) side of the view.
-    /// </summary>
-    Left = 0,
-
-    /// <summary>
-    ///     The top (Y) side of the view.
-    /// </summary>
-    Top = 1,
-
-    /// <summary>
-    ///     The right (X + Width) side of the view.
-    /// </summary>
-    Right = 2,
-
-    /// <summary>
-    ///     The bottom (Y + Height) side of the view.
-    /// </summary>
-    Bottom = 3
-}
 
 /// <summary>
 ///     Describes the position of a <see cref="View"/> which can be an absolute value, a percentage, centered, or
@@ -385,315 +360,4 @@ public abstract class Pos
 
     #endregion operators
 
-}
-
-/// <summary>
-///     Represents an absolute position in the layout. This is used to specify a fixed position in the layout.
-/// </summary>
-/// <remarks>
-///     <para>
-///         This is a low-level API that is typically used internally by the layout system. Use the various static
-///         methods on the <see cref="Pos"/> class to create <see cref="Pos"/> objects instead.
-///     </para>
-/// </remarks>
-/// <param name="position"></param>
-public class PosAbsolute (int position) : Pos
-{
-    /// <summary>
-    ///     The position of the <see cref="View"/> in the layout.
-    /// </summary>
-    public int Position { get; } = position;
-
-    /// <inheritdoc/>
-    public override bool Equals (object other) { return other is PosAbsolute abs && abs.Position == Position; }
-
-    /// <inheritdoc/>
-    public override int GetHashCode () { return Position.GetHashCode (); }
-
-    /// <inheritdoc/>
-    public override string ToString () { return $"Absolute({Position})"; }
-
-    internal override int GetAnchor (int size) { return Position; }
-}
-
-/// <summary>
-///     Represents a position anchored to the end (right side or bottom).
-/// </summary>
-/// <remarks>
-///     <para>
-///         This is a low-level API that is typically used internally by the layout system. Use the various static
-///         methods on the <see cref="Pos"/> class to create <see cref="Pos"/> objects instead.
-///     </para>
-/// </remarks>
-public class PosAnchorEnd : Pos
-{
-    /// <summary>
-    ///     Gets the offset of the position from the right/bottom.
-    /// </summary>
-    public int Offset { get; }
-
-    /// <summary>
-    ///     Constructs a new position anchored to the end (right side or bottom) of the SuperView,
-    ///     minus the respective dimension of the View. This is equivalent to using <see cref="PosAnchorEnd(int)"/>,
-    ///     with an offset equivalent to the View's respective dimension.
-    /// </summary>
-    public PosAnchorEnd () { UseDimForOffset = true; }
-
-    /// <summary>
-    ///     Constructs a new position anchored to the end (right side or bottom) of the SuperView,
-    /// </summary>
-    /// <param name="offset"></param>
-    public PosAnchorEnd (int offset) { Offset = offset; }
-
-    /// <inheritdoc/>
-    public override bool Equals (object other) { return other is PosAnchorEnd anchorEnd && anchorEnd.Offset == Offset; }
-
-    /// <inheritdoc/>
-    public override int GetHashCode () { return Offset.GetHashCode (); }
-
-    /// <summary>
-    ///     If true, the offset is the width of the view, if false, the offset is the offset value.
-    /// </summary>
-    public bool UseDimForOffset { get; }
-
-    /// <inheritdoc/>
-    public override string ToString () { return UseDimForOffset ? "AnchorEnd()" : $"AnchorEnd({Offset})"; }
-
-    internal override int GetAnchor (int size)
-    {
-        if (UseDimForOffset)
-        {
-            return size;
-        }
-
-        return size - Offset;
-    }
-
-    internal override int Calculate (int superviewDimension, Dim dim, View us, Dimension dimension)
-    {
-        int newLocation = GetAnchor (superviewDimension);
-
-        if (UseDimForOffset)
-        {
-            newLocation -= dim.GetAnchor (superviewDimension);
-        }
-
-        return newLocation;
-    }
-}
-
-/// <summary>
-///     Represents a position that is centered.
-/// </summary>
-public class PosCenter : Pos
-{
-    /// <inheritdoc/>
-    public override string ToString () { return "Center"; }
-
-    internal override int GetAnchor (int size) { return size / 2; }
-
-    internal override int Calculate (int superviewDimension, Dim dim, View us, Dimension dimension)
-    {
-        int newDimension = Math.Max (dim.Calculate (0, superviewDimension, us, dimension), 0);
-
-        return GetAnchor (superviewDimension - newDimension);
-    }
-}
-
-/// <summary>
-///     Represents a position that is a combination of two other positions.
-/// </summary>
-/// <remarks>
-///     <para>
-///         This is a low-level API that is typically used internally by the layout system. Use the various static
-///         methods on the <see cref="Pos"/> class to create <see cref="Pos"/> objects instead.
-///     </para>
-/// </remarks>
-/// <param name="add">
-///     Indicates whether the two positions are added or subtracted.
-/// </param>
-/// <param name="left">The left position.</param>
-/// <param name="right">The right position.</param>
-public class PosCombine (AddOrSubtract add, Pos left, Pos right) : Pos
-{
-    /// <summary>
-    ///     Gets whether the two positions are added or subtracted.
-    /// </summary>
-    public AddOrSubtract Add { get; } = add;
-
-    /// <summary>
-    ///     Gets the left position.
-    /// </summary>
-    public new Pos Left { get; } = left;
-
-    /// <summary>
-    ///     Gets the right position.
-    /// </summary>
-    public new Pos Right { get; } = right;
-
-    /// <inheritdoc/>
-    public override string ToString () { return $"Combine({Left}{(Add == AddOrSubtract.Add ? '+' : '-')}{Right})"; }
-
-    internal override int GetAnchor (int size)
-    {
-        int la = Left.GetAnchor (size);
-        int ra = Right.GetAnchor (size);
-
-        if (Add == AddOrSubtract.Add)
-        {
-            return la + ra;
-        }
-
-        return la - ra;
-    }
-
-    internal override int Calculate (int superviewDimension, Dim dim, View us, Dimension dimension)
-    {
-        int left = Left.Calculate (superviewDimension, dim, us, dimension);
-        int right = Right.Calculate (superviewDimension, dim, us, dimension);
-
-        if (Add == AddOrSubtract.Add)
-        {
-            return left + right;
-        }
-
-        return left - right;
-    }
-
-    internal override bool ReferencesOtherViews ()
-    {
-        if (Left.ReferencesOtherViews ())
-        {
-            return true;
-        }
-
-        if (Right.ReferencesOtherViews ())
-        {
-            return true;
-        }
-
-        return false;
-    }
-}
-
-/// <summary>
-///     Represents a position that is a percentage of the width or height of the SuperView.
-/// </summary>
-/// <remarks>
-///     <para>
-///         This is a low-level API that is typically used internally by the layout system. Use the various static
-///         methods on the <see cref="Pos"/> class to create <see cref="Pos"/> objects instead.
-///     </para>
-/// </remarks>
-/// <param name="percent"></param>
-public class PosPercent (float percent) : Pos
-{
-    /// <summary>
-    ///     Gets the factor that represents the percentage of the width or height of the SuperView.
-    /// </summary>
-    public new float Percent { get; } = percent;
-
-    /// <inheritdoc/>
-    public override bool Equals (object other) { return other is PosPercent f && f.Percent == Percent; }
-
-    /// <inheritdoc/>
-    public override int GetHashCode () { return Percent.GetHashCode (); }
-
-    /// <inheritdoc/>
-    public override string ToString () { return $"Percent({Percent})"; }
-
-    internal override int GetAnchor (int size) { return (int)(size * Percent); }
-}
-
-/// <summary>
-///     Represents a position that is computed by executing a function that returns an integer position.
-/// </summary>
-/// <remarks>
-///     <para>
-///         This is a low-level API that is typically used internally by the layout system. Use the various static
-///         methods on the <see cref="Pos"/> class to create <see cref="Pos"/> objects instead.
-///     </para>
-/// </remarks>
-/// <param name="pos">The position.</param>
-public class PosFunc (Func<int> pos) : Pos
-{
-    /// <summary>
-    ///     Gets the function that computes the position.
-    /// </summary>
-    public new Func<int> Func { get; } = pos;
-
-    /// <inheritdoc/>
-    public override bool Equals (object other) { return other is PosFunc f && f.Func () == Func (); }
-
-    /// <inheritdoc/>
-    public override int GetHashCode () { return Func.GetHashCode (); }
-
-    /// <inheritdoc/>
-    public override string ToString () { return $"PosFunc({Func ()})"; }
-
-    internal override int GetAnchor (int size) { return Func (); }
-}
-
-/// <summary>
-///     Represents a position that is anchored to the side of another view.
-/// </summary>
-/// <remarks>
-///     <para>
-///         This is a low-level API that is typically used internally by the layout system. Use the various static
-///         methods on the <see cref="Pos"/> class to create <see cref="Pos"/> objects instead.
-///     </para>
-/// </remarks>
-/// <param name="view">The View the position is anchored to.</param>
-/// <param name="side">The side of the View the position is anchored to.</param>
-public class PosView (View view, Side side) : Pos
-{
-    /// <summary>
-    ///     Gets the View the position is anchored to.
-    /// </summary>
-    public View Target { get; } = view;
-
-    /// <summary>
-    ///     Gets the side of the View the position is anchored to.
-    /// </summary>
-    public Side Side { get; } = side;
-
-    /// <inheritdoc/>
-    public override bool Equals (object other) { return other is PosView abs && abs.Target == Target && abs.Side == Side; }
-
-    /// <inheritdoc/>
-    public override int GetHashCode () { return Target.GetHashCode (); }
-
-    /// <inheritdoc/>
-    public override string ToString ()
-    {
-        string sideString = Side switch
-                            {
-                                Side.Left => "left",
-                                Side.Top => "top",
-                                Side.Right => "right",
-                                Side.Bottom => "bottom",
-                                _ => "unknown"
-                            };
-
-        if (Target == null)
-        {
-            throw new NullReferenceException (nameof (Target));
-        }
-
-        return $"View(side={sideString},target={Target})";
-    }
-
-    internal override int GetAnchor (int size)
-    {
-        return Side switch
-               {
-                   Side.Left => Target.Frame.X,
-                   Side.Top => Target.Frame.Y,
-                   Side.Right => Target.Frame.Right,
-                   Side.Bottom => Target.Frame.Bottom,
-                   _ => 0
-               };
-    }
-
-    internal override bool ReferencesOtherViews () { return true; }
 }
